requirements.txt
discord.py~=2.6.4
aiosqlite~=0.22.1
d20>=1.1.2
